+++ qoffer-domain/src/main/groovy/life/qbic/portal/portlet/customers/create/CreateCustomerOutput.groovy
@@ -1,0 +1,15 @@
+package life.qbic.portal.portlet.customers.create
+
+import life.qbic.portal.portlet.UseCaseFailure
+
+/**
+ * Output interface for the {@link life.qbic.portal.portlet.customers.create.CreateCustomer} use
+ * case
+ *
+ * @since: 1.0.0
+ * @author: Tobias Koch
+ */
+interface CreateCustomerOutput extends UseCaseFailure {
+
+    void customerCreated(String message)
+}