+++ qoffer-domain/src/main/groovy/life/qbic/portal/portlet/customers/update/UpdateCustomer.groovy
@@ -1,0 +1,10 @@
+package life.qbic.portal.portlet.customers.update
+
+/**
+ * A project manager wants to modify the customer information
+ *
+ * @since: 1.0.0
+ * @author: Tobias Koch
+ */
+class UpdateCustomer {
+}