+++ qoffer-domain/src/test/groovy/life/qbic/portal/portlet/customers/update/UpdateCustomerInputSpec.groovy
@@ -1,0 +1,14 @@
+package life.qbic.portal.portlet.customers.update
+
+import spock.lang.Specification
+
+/**
+ * <short description>
+ *
+ * <detailed description>
+ *
+ * @since: 1.0.0
+ * @author: Tobias Koch
+ */
+class UpdateCustomerInputSpec extends Specification {
+}