+++ qoffer-infrastructure/src/test/groovy/life/qbic/portal/qoffer2/database/DatabaseSessionSpec.groovy
@@ -1,0 +1,18 @@
+package life.qbic.portal.qoffer2.database
+
+
+import spock.lang.Specification
+
+/**
+ * Test the database connection
+ *
+ * This class should be used whenever the database connection should be tested
+ *
+ * @since: 1.0.0
+ * @author: Jennifer Bödker
+ *
+ */
+class DatabaseSessionSpec extends Specification{
+
+    //TODO
+}